feat : 댓글 좋아요 등록/삭제

+++ post-service/src/main/java/com/todaytrend/postservice/post/repository/PostLikeRepository.java
@@ -16,5 +16,5 @@ public interface PostLikeRepository extends JpaRepository<PostLike,Long>{
 
     void deleteByUserUuidAndPostId(String userUuid, Long postId);
 
-    Long countByPostId(Long postId);
+    int countByPostId(Long postId);
 }
